Refactored test suites to be more modular
