+++ brambl-sdk/src/test/scala/co/topl/brambl/builders/TransactionBuilderInterpreterSeriesMintingSpec.scala
@@ -1,0 +1,149 @@
+package co.topl.brambl.builders
+
+import co.topl.brambl.builders.TransactionBuilderApi.UnableToBuildTransaction
+import co.topl.brambl.models.Event.SeriesPolicy
+import co.topl.brambl.models.box.{Attestation, Value}
+import co.topl.brambl.models.transaction.{IoTransaction, SpentTransactionOutput, UnspentTransactionOutput}
+import co.topl.brambl.models.Datum
+import co.topl.brambl.syntax.{ioTransactionAsTransactionSyntaxOps, seriesPolicyAsSeriesPolicySyntaxOps}
+import com.google.protobuf.ByteString
+import quivr.models.{Int128, Proof}
+
+class TransactionBuilderInterpreterSeriesMintingSpec extends TransactionBuilderInterpreterSpecBase {
+
+  test("buildSimpleSeriesMintingTransaction > Success") {
+    val mockSeriesPolicy: SeriesPolicy =
+      SeriesPolicy("Mock Series Policy", None, dummyTxoAddress)
+    val quantity = Int128(ByteString.copyFrom(BigInt(1).toByteArray))
+    val expectedTx = IoTransaction.defaultInstance
+      .withDatum(txDatum)
+      .withSeriesPolicies(Seq(Datum.SeriesPolicy(mockSeriesPolicy)))
+      .withInputs(
+        List(
+          SpentTransactionOutput(
+            mockSeriesPolicy.registrationUtxo,
+            Attestation().withPredicate(Attestation.Predicate(inPredicateLockFull, List(Proof()))),
+            Value.defaultInstance.withLvl(Value.LVL(quantity))
+          )
+        )
+      )
+      .withOutputs(
+        List(
+          UnspentTransactionOutput(
+            inLockFullAddress,
+            Value.defaultInstance.withSeries(
+              Value.Series(
+                seriesId = mockSeriesPolicy.computeId,
+                quantity = quantity,
+                tokenSupply = mockSeriesPolicy.tokenSupply,
+                quantityDescriptor = mockSeriesPolicy.quantityDescriptor,
+                fungibility = mockSeriesPolicy.fungibility
+              )
+            )
+          )
+        )
+      )
+    val txRes = txBuilder.buildSimpleSeriesMintingTransaction(
+      inputTxo,
+      inPredicateLockFull,
+      mockSeriesPolicy,
+      quantity,
+      inLockFullAddress
+    )
+    assert(txRes.isRight && txRes.toOption.get.computeId == expectedTx.computeId)
+  }
+
+  test("buildSimpleSeriesMintingTransaction > invalid registrationTxo") {
+    val mockSeriesPolicy: SeriesPolicy =
+      SeriesPolicy("Mock Series Policy", None, dummyTxoAddress.copy(network = 10))
+    val quantity = Int128(ByteString.copyFrom(BigInt(1).toByteArray))
+
+    val testTx = txBuilder
+      .buildSimpleSeriesMintingTransaction(
+        inputTxo,
+        inPredicateLockFull,
+        mockSeriesPolicy,
+        quantity,
+        inLockFullAddress
+      )
+    assertEquals(
+      testTx,
+      Left(
+        UnableToBuildTransaction(
+          Seq(UserInputError("registrationTxo does not match registrationUtxo"))
+        )
+      )
+    )
+  }
+
+  test("buildSimpleSeriesMintingTransaction > invalid registrationUtxo") {
+    val mockSeriesPolicy: SeriesPolicy =
+      SeriesPolicy("Mock Series Policy", None, dummyTxoAddress)
+    val quantity = Int128(ByteString.copyFrom(BigInt(1).toByteArray))
+
+    val testTx = txBuilder
+      .buildSimpleSeriesMintingTransaction(
+        inputTxo.copy(transactionOutput =
+          fullOutput.copy(value = Value.defaultInstance.withTopl(Value.TOPL(quantity)))
+        ),
+        inPredicateLockFull,
+        mockSeriesPolicy,
+        quantity,
+        inLockFullAddress
+      )
+    assertEquals(
+      testTx,
+      Left(
+        UnableToBuildTransaction(
+          Seq(UserInputError("registrationUtxo does not contain LVLs"))
+        )
+      )
+    )
+  }
+
+  test("buildSimpleSeriesMintingTransaction > invalid registrationLock") {
+    val mockSeriesPolicy: SeriesPolicy =
+      SeriesPolicy("Mock Series Policy", None, dummyTxoAddress)
+    val quantity = Int128(ByteString.copyFrom(BigInt(1).toByteArray))
+
+    val testTx = txBuilder
+      .buildSimpleSeriesMintingTransaction(
+        inputTxo,
+        trivialOutLock.getPredicate,
+        mockSeriesPolicy,
+        quantity,
+        inLockFullAddress
+      )
+    assertEquals(
+      testTx,
+      Left(
+        UnableToBuildTransaction(
+          Seq(UserInputError("registrationLock does not correspond to registrationTxo"))
+        )
+      )
+    )
+  }
+
+  test("buildSimpleSeriesMintingTransaction > invalid quantityToMint") {
+    val mockSeriesPolicy: SeriesPolicy =
+      SeriesPolicy("Mock Series Policy", None, dummyTxoAddress)
+    val quantity = Int128(ByteString.copyFrom(BigInt(0).toByteArray))
+
+    val testTx = txBuilder
+      .buildSimpleSeriesMintingTransaction(
+        inputTxo,
+        inPredicateLockFull,
+        mockSeriesPolicy,
+        quantity,
+        inLockFullAddress
+      )
+    assertEquals(
+      testTx,
+      Left(
+        UnableToBuildTransaction(
+          Seq(UserInputError("quantityToMint must be positive"))
+        )
+      )
+    )
+  }
+}
